fix entidades ,dtos , enum y mappers:  -variable tipado intensidad de entrenamiento  - builder  - exerciseInfo

+++ src/main/java/com/jpmt/strengthlab/models/dto/workoutset/WorkoutSetResponse.java
@@ -11,7 +11,7 @@ public record WorkoutSetResponse(
         Integer reps,
         Double weight,
         IntensityType intensityType,
-        Integer intensityValue,
+        Double intensityValue,
         OffsetDateTime createdAt,
         OffsetDateTime updatedAt
 ) {
